==> sider/src/01_SIDER_compound_mapping.ipynb <==
- Remove zeros in SIDER STITCH IDs to get PubChem Compound ID
- Use Unichem to get InchiKey for PubChem ID
- also get chembl_id
- do query for parent molregno in ChEMBL

import pandas as pd
import requests
import re
import json
import time
import pickle
import datetime

current_date = datetime.date.today().strftime("%Y%m%d")

pd.set_option('display.max_rows', 1000)
pd.set_option('display.max_colwidth', 30)

basedir = "/scratch/ias41/ae_code"

### Whole of SIDER

all_se = pd.read_csv(basedir + '/sider/data/meddra_all_se.tsv', sep='\t', header=None)
all_se.columns = ['STITCH_flat', 'STITCH_stereo', 'UMLS_concept_label', 'MedDRA_concept_type','UMLS_concept', 'side_effect_name']

all_se.head()

len(set(all_se['STITCH_stereo']))

### Convert STITCH ids to InChIs

p = re.compile('CID0*([0-9]+)')
base = 'https://www.ebi.ac.uk/unichem/rest/'

inchi_dict = dict()
errors = []

def find_inchi(stitch_id):
    """For a STITCH compound id, retrieve the inchi and inchikey via Unichem
    stitch_id -- str, STITCH compound id, which has arbitrary number of 0s inserted between the 'CID' and the PubChem id."""
    
    compound_id = p.match(stitch_id).group(1)
    response = requests.get(base + 'src_compound_id/{}/22/1'.format(compound_id)) # source 22 is PubChem
    
    try:
        assert response.status_code == 200
    except AssertionError:
        errors.append(stitch_id)
        return
    
    result = json.loads(response.content.decode())
    
    try:
        chembl_id = result[0]['src_compound_id']
    except IndexError:
        errors.append(stitch_id)
        return
    time.sleep(0.2)
    
    inchi_dict['CID' + compound_id] = {'original_id': stitch_id, 'chembl_id': chembl_id}

%%time
for compound in all_se['STITCH_stereo'].drop_duplicates():
    find_inchi(compound)

len(inchi_dict)

# Many of the errors are biologicals/peptides or are compounds that have been downgraded - checked some manually
len(errors)

339/1556

with open(basedir + '/sider/data/stitch2inchi_dict.pkl', 'wb') as f:
    pickle.dump(inchi_dict, f)

with open(basedir + '/sider/data/stitch2inchi_dict.pkl', 'rb') as f:
    inchi_from_pkl = pickle.load(f)

### Make mapping dataframe

compound_mappings = pd.DataFrame.from_dict(inchi_dict, orient='index').reset_index()
compound_mappings.columns = ['pubchem_id', 'stitch_id', 'chembl_id']

compound_mappings.head()

### Find parent compounds in ChEMBL

len(compound_mappings['stitch_id'].drop_duplicates())

len(compound_mappings['chembl_id'].drop_duplicates())

myquery = """use chembl_25;
select distinct
md.chembl_id
, md.molregno
, hier.parent_molregno
, md2.chembl_id as parent_chembl_id
from molecule_dictionary md
left join molecule_hierarchy hier on hier.molregno = md.molregno
left join molecule_dictionary md2 on md2.molregno = hier.parent_molregno
where md.chembl_id in ({});""".format("'" + "', '".join(list(compound_mappings['chembl_id'])) + "'")

with open(basedir + '/sider/src/sider_parent_molregnos.sql', 'w') as f:
    f.write(myquery)

#### Executed query on local MySQL installation of ChEMBL

parent_results = pd.read_csv(basedir + '/sider/data/sider_parent_molregnos.txt', sep='\t')

# Sometimes parent molregnos are missing - replace by molregno

parent_results['parent_chembl_id'] = parent_results.parent_chembl_id.fillna(value=parent_results.chembl_id)

parent_results['parent_molregno'] = parent_results.parent_molregno.fillna(value=parent_results.molregno)

parent_results['parent_molregno'] = parent_results['parent_molregno'].astype(int)

parent_results.head()

len(parent_results['chembl_id'].drop_duplicates())

len(parent_results['parent_molregno'].drop_duplicates())

# Need to add back original identifiers 

# reverse the inchi dict
chembl_id_dict = dict()

for key in inchi_dict.keys():
    chembl_id = inchi_dict[key]['chembl_id']
    original_id = inchi_dict[key]['original_id']
    pubchem_id = key
    chembl_id_dict[chembl_id] = {'stitch_id': original_id, 'pubchem_id': key}

chembl_id_dict

parent_results['stitch_stereo'] = parent_results['chembl_id'].apply(lambda x: chembl_id_dict[x]['stitch_id'])

parent_results.head()

# Just check the parent query has worked
parent_results.loc[parent_results['parent_molregno']!=parent_results['molregno'],:].head()

# check if any were null
parent_results.loc[parent_results['parent_chembl_id'].isnull()]

# Save convenience file with mappings stitch 2 parent chembl molregno
parent_results.to_csv(basedir + '/sider/data/sider_mapped_stitch2parent.txt', sep='\t', index=None)

pt_mapped = all_se.merge(parent_results, left_on='STITCH_stereo', right_on='stitch_stereo')
pt_mapped = pt_mapped.loc[pt_mapped['MedDRA_concept_type']=='PT',:]

pt_mapped.head()

# Save version of all_se file for compounds with molregno mappings
pt_mapped.to_csv(basedir + '/sider/results/sider_all_se_pt_mapped.txt', sep='\t', index=None)
