==> sider/src/02_sider_make_clinical_dataset.ipynb <==
# SIDER
# Want to select as best as possible, all the SEs that are from clinical trials (as opposed to post-marketing)
# Cannot simply exclude all those SEs marked as pm because e.g. nausea (epi/doxorubucin) is observed in clin trials and pm, and would then be excluded
# 1. If the label has no pm SEs at all > assume all are clinical trial
# 2. If the label has some pm > check which ones are 'unique' to pm, i.e. exclude those pairs which also have a frequency
# because if they have a frequency this is only derived from clinical trials

# use mappings to molregno/chembl_id I did earlier
# save file with clinical trial pairs > use for doing target - adverse event analysis 
# save file with postmarketing pairs > use for doing target -adverse event analysis 


import pandas as pd

sider_dir = '/scratch/ias41/ae_code/sider'

# Load all reported side effects
all_se = pd.read_csv(sider_dir + '/data/meddra_all_se.tsv', sep='\t', header=None)
all_se.columns = ['stitch_flat', 'stitch_stereo', 'umls_concept_label', 'meddra_concept_type','umls_concept', 'side_effect_name']
all_se_pt = all_se.loc[all_se['meddra_concept_type']=='PT',:]

# First, need to exclude some rows from the frequency file which have been reported as wrong (meddra_freq_to_exclude.tsv) from GitHub page
# https://github.com/mkuhn/sider

to_exclude = pd.read_csv(sider_dir + '/data/meddra_freq_to_exclude_tab.tsv', sep='\t',header=None)
to_exclude.columns = ['stitch_flat', 'stitch_stereo', 'umls_concept_label', 'placebo','frequency_description', 'lower_bound', 'upper_bound', 'meddra_concept_type', 'umls_concept', 'side_effect_name']

to_exclude

# Load side effects from post-marketing or with frequency
freq_se = pd.read_csv(sider_dir + '/data/meddra_freq.tsv', sep='\t', header=None)
freq_se.columns = ['stitch_flat', 'stitch_stereo', 'umls_concept_label', 'placebo','frequency_description', 'lower_bound', 'upper_bound', 'meddra_concept_type', 'umls_concept', 'side_effect_name']
freq_se_pt = freq_se.loc[freq_se['meddra_concept_type']=='PT',:]

# first concatenate the 'to exclude' rows, then exclude duplicates, to exclude the rows that are wrong
to_exclude_concat = pd.concat([freq_se, to_exclude])
freq_se_corrected = to_exclude_concat.drop_duplicates(keep=False)

freq_se_corrected_pt = freq_se_corrected.loc[freq_se_corrected['meddra_concept_type']=='PT',:]

# Make selections of postmarketing and freq_rest pairs
pm = freq_se_corrected_pt.loc[freq_se_corrected_pt['frequency_description']=='postmarketing',:]
freq_rest = freq_se_corrected_pt.loc[freq_se_corrected_pt['frequency_description']!='postmarketing',:]

len(pm.drop_duplicates())

#All of SIDER
len(all_se_pt[['stitch_stereo', 'side_effect_name']].drop_duplicates())

# % post-marketing
len(pm[['stitch_stereo', 'side_effect_name']].drop_duplicates())

21209/152759

len(all_se_pt[['stitch_stereo']].drop_duplicates())

# Next 
# 1. If the label has no pm SEs at all > assume all are clinical trial

# Find the drugs that do not have any postmarketing events
pm_drugs = set(pm['stitch_stereo'])

# This dataframe contains all drugs+SEs that we will accept as clinical trial because they have no postmarketing SEs
no_pm_selection = all_se_pt.loc[~all_se_pt['stitch_stereo'].isin(pm_drugs),:]

# check some counts
len(pm_drugs)

len(set(no_pm_selection['stitch_stereo']))

len(pm_drugs) + len(set(no_pm_selection['stitch_stereo']))

len(set(all_se_pt['stitch_stereo']))

928 drugs (out of total 1556 in all_se SIDER) did not have any postmarketing effects, so they were assumed as clinical trials

# 2. Find post-marketing events, but exclude those which also have frequencies

# compound-se pairs with frequencies
freq_pairs = set([(item[1]['stitch_stereo'],item[1]['side_effect_name']) for item in freq_rest.iterrows()])

# Divide the rows from pm into those with are 'unique' to PM and those which also have a frequency

pm_unique = []
pm_and_freq = []

for item in pm.iterrows():
    if (item[1]['stitch_stereo'],item[1]['side_effect_name']) in freq_pairs:
        pm_and_freq.append(item[1])
    elif (item[1]['stitch_stereo'],item[1]['side_effect_name']) not in freq_pairs:
        pm_unique.append(item[1])

len(pm)

len(pm_unique) + len(pm_and_freq)

# Of all drug-se effect pairs which were marked as pm:
# Fraction which also had frequency (assigned to clinical trial):
len(pm_and_freq) / len(pm)

# This leave this fraction of drug-se pairs which were uniquely marked as pm
len(pm_unique) / len(pm)

pm_unique_df = pd.DataFrame(pm_unique)
pm_and_freq_df = pd.DataFrame(pm_and_freq)

# we will accept the drug-se pairs in pm_and_freq as also clinical trial findings
# because frequencies are normally found in clinical trials

clin_trial = pd.concat([no_pm_selection, pm_and_freq_df], sort=False)

# There should not be any drug-ae pairs overlapping between the clinical trial and the pm_unique ... 

pm_unique_pairs = set([(item[1]['stitch_stereo'],item[1]['side_effect_name']) for item in pm_unique_df.iterrows()])

clin_trial_pairs = set([(item[1]['stitch_stereo'],item[1]['side_effect_name']) for item in clin_trial.iterrows()])

len(pm_unique_pairs & clin_trial_pairs)

pm_and_freq_pairs = set([(item[1]['stitch_stereo'],item[1]['side_effect_name']) for item in pm_and_freq_df.iterrows()])

# How many drug-se pairs, unique drugs, and unique SEs in the two new files?
print('Post-marketing specific events:')
print('Number of unique drug-se pairs: {}'.format(len(pm_unique_pairs)))
print('Number of unique drugs: {}'.format(len(set(pm_unique_df['stitch_stereo']))))
print('Number of unique events: {}'.format(len(set(pm_unique_df['side_effect_name']))))

print('\nEvents in both postmarketing and with frequency:')
print('Number of unique drug-se pairs: {}'.format(len(pm_and_freq_pairs)))
print('Number of unique drugs: {}'.format(len(set(pm_and_freq_df['stitch_stereo']))))
print('Number of unique events: {}'.format(len(set(pm_and_freq_df['side_effect_name']))))

print('\nEvents of drugs without any postmarketing data:')
print('Number of unique drug-se pairs: {}'.format(len(set([(item[1]['stitch_stereo'],item[1]['side_effect_name']) for item in no_pm_selection.iterrows()]))))
print('Number of unique drugs: {}'.format(len(set(no_pm_selection['stitch_stereo']))))
print('Number of unique events: {}'.format(len(set(no_pm_selection['side_effect_name']))))

print('\nEvents assigned to clinical trials set:')
print('Number of unique drug-se pairs: {}'.format(len(clin_trial_pairs)))
print('Number of unique drugs: {}'.format(len(set(clin_trial['stitch_stereo']))))
print('Number of unique events: {}'.format(len(set(clin_trial['side_effect_name']))))

# Attach chembl identifiers to the IDs

len(set(pm_unique_df['stitch_stereo']))

# Load compound mappings
parent_results = pd.read_csv(sider_dir + '/data/sider_mapped_stitch2parent.txt', sep='\t')

parent_results.head()

# Merge pm-specific pairs with the compound mappings
pm_unique_df_mapped = pm_unique_df.merge(parent_results, on='stitch_stereo')

# Drug-event pairs assigned to postmarketing set - only those with mapped compounds to ChEMBL
#pm_unique_df_mapped.to_csv(sider_dir + '/results/sider_processed_pm-assigned_mapped.txt', sep='\t', index=None)

# 500 out of 619 compounds for the postmarketing-specific drug-se pairs were mapped
len(set(pm_unique_df_mapped['stitch_stereo']))

# Merge clin_trial-specific pairs with the compound mappings

clin_trial.head()

len(set(clin_trial['stitch_stereo']))

clin_trial_mapped = clin_trial.merge(parent_results, on='stitch_stereo')

len(set(clin_trial_mapped['stitch_stereo']))

# Drug-event pairs assigned to clinical trials set - only those with mapped compounds to ChEMBL
clin_trial_mapped.to_csv(sider_dir + '/results/sider_processed_clin_trial-assigned_mapped.txt', sep='\t', index=None)

# 1041 out of 1333 compounds for the clinical-trial assigned drug-se pairs were mapped to compounds

clin_trial_mapped.head()
