==> sider/src/03_make_ae_pickle.ipynb <==
import pandas as pd
import pickle
import collections
import datetime

sider_dir = '/scratch/ias41/ae_code/sider'

report_information = collections.OrderedDict()

# Drug-event pairs assigned to clinical trials set - only those with mapped compounds to ChEMBL
clin_trial_mapped = pd.read_csv(sider_dir + '/results/sider_processed_clin_trial-assigned_mapped.txt', sep='\t')

clin_trial_mapped.head()

# Get some basic counts for the report

# Nr of stereo compounds in sider all_se with Meddra PTs
# Load all reported side effects
all_se = pd.read_csv(sider_dir + '/data/meddra_all_se.tsv', sep='\t', header=None)
all_se.columns = ['stitch_flat', 'stitch_stereo', 'umls_concept_label', 'meddra_concept_type','umls_concept', 'side_effect_name']
all_se_pt = all_se.loc[all_se['meddra_concept_type']=='PT',:]

report_information['Datapoints in all SIDER all_se (before mapping)'] = '{} unique stereo_stich ids, {} unique AEs (PTs)'.format(len(all_se_pt['stitch_stereo'].drop_duplicates()), len(all_se_pt['side_effect_name'].drop_duplicates()))

# Nr of stereo compounds in sider_all_se mapped to molregnos

all_se_pt_mapped = pd.read_csv(sider_dir + '/results/sider_all_se_pt_mapped.txt', sep='\t')

report_information['Datapoints in all SIDER all_se (mapped)'] = '{} unique stereo_stich ids, {} unique AEs (PTs)'.format(len(all_se_pt_mapped['STITCH_stereo'].drop_duplicates()), len(all_se_pt_mapped['side_effect_name'].drop_duplicates()))

# Nr of mapped stereo ids mapped assigned to clinical set

report_information['Datapoints in clinical-assigned SIDER (mapped)'] = '{} unique stereo_stich ids, {} unique AEs (PTs)'.format(len(clin_trial_mapped['stitch_stereo'].drop_duplicates()), len(clin_trial_mapped['side_effect_name'].drop_duplicates()))


report_information

# Find side effects with less than five drugs
ses_with_less_than_five_drugs = list()
for group in clin_trial_mapped.groupby(by='side_effect_name'):
    if len(group[1]['parent_molregno'].drop_duplicates()) < 5:
        ses_with_less_than_five_drugs.append(group[0])

clin_trial_mapped.loc[clin_trial_mapped['side_effect_name']=='Abdominal hernia',:]

len(ses_with_less_than_five_drugs)

report_information['Number of side effects (from clinical mapped) excluded because they had fewer than 5 drugs'] = len(ses_with_less_than_five_drugs)

# Exclude SEs with less than five drugs
sider_data_selected = clin_trial_mapped.loc[~clin_trial_mapped['side_effect_name'].isin(ses_with_less_than_five_drugs),:]

report_information['Datapoints in clinical-assigned SIDER (mapped) after excluding side effects without 5 drugs (used to make ae pickle)'] = '{} unique stereo_stich ids, {} unique AEs (PTs)'.format(len(sider_data_selected['stitch_stereo'].drop_duplicates()), len(sider_data_selected['side_effect_name'].drop_duplicates()))

report_information

molregno2aes_sider = dict()

for group in sider_data_selected.groupby(by='parent_molregno'):
    molregno = group[0]
    aes = list(group[1]['side_effect_name'].drop_duplicates())
    
    molregno2aes_sider[molregno] = set()
    for ae in aes:
        molregno2aes_sider[molregno].add(ae.upper())

current_date = datetime.date.today().strftime("%Y%m%d")
with open(sider_dir + '/results/{}_molregno2aes_sider_min5drugs.pkl'.format(current_date), 'wb') as f:
    pickle.dump(molregno2aes_sider, f)

# Write report with counts
with open(sider_dir + '/results/report_counts_sider_clinical_set.txt', 'w') as f:
    for k, v in report_information.items():
        f.write(k + ': ' + str(v) + '\n')
